#474 Move some PSyIR nodes tests from psyGen_test.py to tests/psyir/nodes/*
