+++ src/psyclone/tests/psyGen_test.py
@@ -70,7 +70,7 @@ from psyclone.transformations import OMPParallelLoopTrans, \
     ACCEnterDataTrans, ACCParallelTrans, ACCLoopTrans, ACCKernelsTrans
 from psyclone.generator import generate
 from psyclone.configuration import Config
-from psyclone.tests.utilities import get_invoke
+from psyclone.tests.utilities import get_invoke, check_links
 
 BASE_PATH = os.path.join(os.path.dirname(os.path.abspath(__file__)),
                          "test_files", "dynamo0p3")
@@ -86,31 +86,7 @@ def setup():
     Config.get().api = "dynamo0.3"
 
 
-# Utility functions
-
-def check_links(parent, children):
-    '''Utilitiy routine to check that the parent node has children as its
-    children in the order specified and that the children have parent
-    as their parent. Also check that the parent does not have any
-    additional children that are not provided in the children
-    argument.
-
-    :param parent: the parent node that should have the child \
-        nodes as its children.
-    :type parent: :py:class:`psyclone.psyGen.Node`
-    :param children: the child nodes that should have the parent \
-        node as their parent.
-    :type parent: list of :py:class:`psyclone.psyGen.Node`
-
-    '''
-    assert len(parent.children) == len(children)
-    for index, child in enumerate(children):
-        assert child.parent is parent
-        assert parent.children[index] is child
-
-
 # Tests for utilities
-
 
 def test_object_index():
     ''' Tests for the object_index() utility. '''
@@ -571,54 +547,6 @@ end module dummy_mod
 '''
 
 
-# Schedule class tests
-
-def test_sched_node_str():
-    ''' Check the node_str method of the Schedule class'''
-    from psyclone.psyir.nodes.node import colored, SCHEDULE_COLOUR_MAP
-    sched = Schedule()
-    assert colored("Schedule", SCHEDULE_COLOUR_MAP["Schedule"]) in \
-        sched.node_str()
-
-
-def test_sched_getitem():
-    '''Test that Schedule has the [int] operator overloaded to return the
-    given index child'''
-    _, invoke_info = parse(os.path.join(BASE_PATH,
-                                        "15.9.1_X_innerproduct_Y_builtin.f90"),
-                           api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3", distributed_memory=True).create(invoke_info)
-
-    sched = psy.invokes.invoke_list[0].schedule
-    for indx in range(len(sched._children)):
-        assert sched[indx] is sched._children[indx]
-
-    # Test range indexing
-    children = sched[:]
-    assert len(children) == 2
-    assert children[0] is sched._children[0]
-    assert children[1] is sched._children[1]
-
-    # Test index out-of-bounds Error
-    with pytest.raises(IndexError) as err:
-        _ = sched[len(sched._children)]
-    assert "list index out of range" in str(err.value)
-
-
-def test_sched_can_be_printed():
-    ''' Check the schedule class can always be printed'''
-    _, invoke_info = parse(os.path.join(BASE_PATH,
-                                        "15.9.1_X_innerproduct_Y_builtin.f90"),
-                           api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3", distributed_memory=True).create(invoke_info)
-
-    # For this test use the generic class
-    psy.invokes.invoke_list[0].schedule.__class__ = Schedule
-    output = str(psy.invokes.invoke_list[0].schedule)
-
-    assert "Schedule:\n" in output
-
-
 # InvokeSchedule class tests
 
 def test_invokeschedule_node_str():
@@ -663,72 +591,6 @@ def test_invokeschedule_can_be_printed():
     output = str(psy.invokes.invoke_list[0].schedule)
 
     assert "InvokeSchedule:\n" in output
-
-
-# Loop class test
-def test_loop_create():
-    '''Test that the create method in the Loop class correctly
-    creates a Loop instance.
-
-    '''
-    start = Literal("0", DataType.INTEGER)
-    stop = Literal("1", DataType.INTEGER)
-    step = Literal("1", DataType.INTEGER)
-    child_node = Assignment.create(Reference("tmp"), Reference("i"))
-    loop = Loop.create("i", start, stop, step, [child_node])
-    schedule = loop.children[3]
-    assert isinstance(schedule, Schedule)
-    check_links(loop, [start, stop, step, schedule])
-    check_links(schedule, [child_node])
-    result = FortranWriter().loop_node(loop)
-    assert result == "do i = 0, 1, 1\n  tmp=i\nenddo\n"
-
-
-def test_loop_create_invalid():
-    '''Test that the create method in a Loop class raises the expected
-    exception if the provided input is invalid.
-
-    '''
-    zero = Literal("0", DataType.INTEGER)
-    one = Literal("1", DataType.INTEGER)
-    children = [Assignment.create(Reference("x"), one)]
-
-    # var_name is not a string.
-    with pytest.raises(GenerationError) as excinfo:
-        _ = Loop.create(1, zero, one, one, children)
-    assert ("var_name argument in create method of Loop class "
-            "should be a string but found 'int'.") in str(excinfo.value)
-
-    # start not a Node.
-    with pytest.raises(GenerationError) as excinfo:
-        _ = Loop.create("i", "invalid", one, one, children)
-    assert ("start argument in create method of Loop class should "
-            "be a PSyIR Node but found 'str'.") in str(excinfo.value)
-
-    # stop not a Node.
-    with pytest.raises(GenerationError) as excinfo:
-        _ = Loop.create("i", zero, "invalid", one, children)
-    assert ("stop argument in create method of Loop class should "
-            "be a PSyIR Node but found 'str'.") in str(excinfo.value)
-
-    # step not a Node.
-    with pytest.raises(GenerationError) as excinfo:
-        _ = Loop.create("i", zero, one, "invalid", children)
-    assert ("step argument in create method of Loop class should "
-            "be a PSyIR Node but found 'str'.") in str(excinfo.value)
-
-    # children not a list
-    with pytest.raises(GenerationError) as excinfo:
-        _ = Loop.create("i", zero, one, one, "invalid")
-    assert ("children argument in create method of Loop class should "
-            "be a list but found 'str'." in str(excinfo.value))
-
-    # contents of children list are not Node.
-    with pytest.raises(GenerationError) as excinfo:
-        _ = Loop.create("i", zero, one, one, ["invalid"])
-    assert (
-        "child of children argument in create method of Loop class "
-        "should be a PSyIR Node but found 'str'." in str(excinfo.value))
 
 
 # Kern class test
@@ -1517,135 +1379,6 @@ def test_argument_backward_dependence(monkeypatch, annexed):
     assert result == prev_arg
 
 
-def test_node_depth():
-    '''
-    Test that the Node class depth method returns the correct value for a
-    Node in a tree. The start depth to determine a Node's depth is set to
-    0. Depth of a Schedule is 1 and increases for its descendants.
-    '''
-    _, invoke_info = parse(
-        os.path.join(BASE_PATH, "1_single_invoke.f90"),
-        api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3", distributed_memory=True).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    schedule = invoke.schedule
-    # Assert that start_depth of any Node (including Schedule) is 0
-    assert schedule.START_DEPTH == 0
-    # Assert that Schedule depth is 1
-    assert schedule.depth == 1
-    # Depth increases by 1 for descendants at each level
-    for child in schedule.children:
-        assert child.depth == 2
-    for child in schedule.children[3].children:
-        assert child.depth == 3
-
-
-def test_node_position():
-    '''
-    Test that the Node class position and abs_position methods return
-    the correct value for a Node in a tree. The start position is
-    set to 0. Relative position starts from 0 and absolute from 1.
-    '''
-    _, invoke_info = parse(
-        os.path.join(BASE_PATH, "4.7_multikernel_invokes.f90"),
-        api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3", distributed_memory=True).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    schedule = invoke.schedule
-    child = schedule.children[6]
-    # Assert that position of a Schedule (no parent Node) is 0
-    assert schedule.position == 0
-    # Assert that start_position of any Node is 0
-    assert child.START_POSITION == 0
-    # Assert that relative and absolute positions return correct values
-    assert child.position == 6
-    assert child.abs_position == 7
-    # Test InternalError for _find_position with an incorrect position
-    with pytest.raises(InternalError) as excinfo:
-        _, _ = child._find_position(child.root.children, -2)
-    assert "started from -2 instead of 0" in str(excinfo.value)
-    # Test InternalError for abs_position with a Node that does
-    # not belong to the Schedule
-    ompdir = OMPDoDirective()
-    with pytest.raises(InternalError) as excinfo:
-        _ = ompdir.abs_position
-    assert ("PSyclone internal error: Error in search for Node position "
-            "in the tree") in str(excinfo.value)
-
-
-def test_node_root():
-    '''
-    Test that the Node class root method returns the correct instance
-    for a Node in a tree.
-    '''
-    _, invoke_info = parse(
-        os.path.join(BASE_PATH, "4.7_multikernel_invokes.f90"),
-        api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3", distributed_memory=False).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    ru_schedule = invoke.schedule
-    # Select a loop and the kernel inside
-    ru_loop = ru_schedule.children[1]
-    ru_kern = ru_loop.children[0]
-    # Assert that the absolute root is a Schedule
-    assert isinstance(ru_kern.root, Schedule)
-
-
-def test_node_annotations():
-    '''Test that an instance of the Node class raises an exception if an
-    annotation is invalid. Note, any annotation will be invalid here
-    as Node does not set a list of valid annotations (this is the job
-    of the subclass).
-
-    '''
-    with pytest.raises(InternalError) as excinfo:
-        _ = Node(annotations=["invalid"])
-    assert (
-        "Node with unrecognized annotation 'invalid', valid annotations are: "
-        "()." in str(excinfo.value))
-
-
-def test_node_args():
-    '''Test that the Node class args method returns the correct arguments
-    for Nodes that do not have arguments themselves'''
-    _, invoke_info = parse(
-        os.path.join(BASE_PATH, "4_multikernel_invokes.f90"),
-        api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3", distributed_memory=False).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    schedule = invoke.schedule
-    loop1 = schedule.children[0]
-    kern1 = loop1.loop_body[0]
-    loop2 = schedule.children[1]
-    kern2 = loop2.loop_body[0]
-    # 1) Schedule (not that this is useful)
-    all_args = kern1.arguments.args
-    all_args.extend(kern2.arguments.args)
-    schedule_args = schedule.args
-    for idx, arg in enumerate(all_args):
-        assert arg == schedule_args[idx]
-    # 2) Loop1
-    loop1_args = loop1.args
-    for idx, arg in enumerate(kern1.arguments.args):
-        assert arg == loop1_args[idx]
-    # 3) Loop2
-    loop2_args = loop2.args
-    for idx, arg in enumerate(kern2.arguments.args):
-        assert arg == loop2_args[idx]
-    # 4) Loop fuse
-    ftrans = DynamoLoopFuseTrans()
-    ftrans.same_space = True
-    schedule, _ = ftrans.apply(schedule.children[0], schedule.children[1])
-    loop = schedule.children[0]
-    kern1 = loop.loop_body[0]
-    kern2 = loop.loop_body[1]
-    loop_args = loop.args
-    kern_args = kern1.arguments.args
-    kern_args.extend(kern2.arguments.args)
-    for idx, arg in enumerate(kern_args):
-        assert arg == loop_args[idx]
-
-
 def test_call_args():
     '''Test that the call class args method returns the appropriate
     arguments '''
@@ -1726,120 +1459,6 @@ def test_globalsum_args():
     global_sum = schedule.children[2]
     assert len(global_sum.args) == 1
     assert global_sum.args[0] == global_sum.scalar
-
-
-def test_node_forward_dependence():
-    '''Test that the Node class forward_dependence method returns the
-    closest dependent Node after the current Node in the schedule or
-    None if none are found.'''
-    _, invoke_info = parse(
-        os.path.join(BASE_PATH, "15.14.1_multi_aX_plus_Y_builtin.f90"),
-        api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3", distributed_memory=True).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    schedule = invoke.schedule
-    read4 = schedule.children[4]
-    # 1: returns none if none found
-    # a) check many reads
-    assert not read4.forward_dependence()
-    # b) check no dependencies for a call
-    assert not read4.children[0].forward_dependence()
-    # 2: returns first dependent kernel arg when there are many
-    # dependencies
-    # a) check first read returned
-    writer = schedule.children[3]
-    next_read = schedule.children[4]
-    assert writer.forward_dependence() == next_read
-    # a) check writer returned
-    first_loop = schedule.children[0]
-    assert first_loop.forward_dependence() == writer
-    # 3: haloexchange dependencies
-    _, invoke_info = parse(
-        os.path.join(BASE_PATH, "4.5_multikernel_invokes.f90"),
-        api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3", distributed_memory=True).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    schedule = invoke.schedule
-    prev_loop = schedule.children[7]
-    halo_field = schedule.children[8]
-    next_loop = schedule.children[9]
-    # a) previous loop depends on halo exchange
-    assert prev_loop.forward_dependence() == halo_field
-    # b) halo exchange depends on following loop
-    assert halo_field.forward_dependence() == next_loop
-
-    # 4: globalsum dependencies
-    _, invoke_info = parse(
-        os.path.join(BASE_PATH, "15.14.3_sum_setval_field_builtin.f90"),
-        api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3", distributed_memory=True).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    schedule = invoke.schedule
-    prev_loop = schedule.children[0]
-    sum_loop = schedule.children[1]
-    global_sum_loop = schedule.children[2]
-    next_loop = schedule.children[3]
-    # a) prev loop depends on sum loop
-    assert prev_loop.forward_dependence() == sum_loop
-    # b) sum loop depends on global sum loop
-    assert sum_loop.forward_dependence() == global_sum_loop
-    # c) global sum loop depends on next loop
-    assert global_sum_loop.forward_dependence() == next_loop
-
-
-def test_node_backward_dependence():
-    '''Test that the Node class backward_dependence method returns the
-    closest dependent Node before the current Node in the schedule or
-    None if none are found.'''
-    _, invoke_info = parse(
-        os.path.join(BASE_PATH, "15.14.1_multi_aX_plus_Y_builtin.f90"),
-        api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3", distributed_memory=True).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    schedule = invoke.schedule
-    # 1: loop no backwards dependence
-    loop3 = schedule.children[2]
-    assert not loop3.backward_dependence()
-    # 2: loop to loop backward dependence
-    # a) many steps
-    last_loop_node = schedule.children[6]
-    prev_dep_loop_node = schedule.children[3]
-    assert last_loop_node.backward_dependence() == prev_dep_loop_node
-    # b) previous
-    assert prev_dep_loop_node.backward_dependence() == loop3
-    # 3: haloexchange dependencies
-    _, invoke_info = parse(
-        os.path.join(BASE_PATH, "4.5_multikernel_invokes.f90"),
-        api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3", distributed_memory=True).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    schedule = invoke.schedule
-    loop2 = schedule.children[7]
-    halo_exchange = schedule.children[8]
-    loop3 = schedule.children[9]
-    # a) following loop node depends on halo exchange node
-    result = loop3.backward_dependence()
-    assert result == halo_exchange
-    # b) halo exchange node depends on previous loop node
-    result = halo_exchange.backward_dependence()
-    assert result == loop2
-    # 4: globalsum dependencies
-    _, invoke_info = parse(
-        os.path.join(BASE_PATH, "15.14.3_sum_setval_field_builtin.f90"),
-        api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3", distributed_memory=True).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    schedule = invoke.schedule
-    loop1 = schedule.children[0]
-    loop2 = schedule.children[1]
-    global_sum = schedule.children[2]
-    loop3 = schedule.children[3]
-    # a) loop3 depends on global sum
-    assert loop3.backward_dependence() == global_sum
-    # b) global sum depends on loop2
-    assert global_sum.backward_dependence() == loop2
-    # c) loop2 (sum) depends on loop1
-    assert loop2.backward_dependence() == loop1
 
 
 def test_call_forward_dependence():
@@ -2023,129 +1642,6 @@ def test_directive_get_private(monkeypatch):
         _ = directive._get_private_list()
     assert ("call 'testkern_code' has a local variable but its name is "
             "not set" in str(err.value))
-
-
-def test_node_is_valid_location():
-    '''Test that the Node class is_valid_location method returns True if
-    the new location does not break any data dependencies, otherwise it
-    returns False'''
-    _, invoke_info = parse(
-        os.path.join(BASE_PATH, "1_single_invoke.f90"),
-        api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3", distributed_memory=True).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    schedule = invoke.schedule
-    # 1: new node argument is invalid
-    node = schedule.children[0]
-    with pytest.raises(GenerationError) as excinfo:
-        node.is_valid_location("invalid_node_argument")
-    assert "argument is not a Node, it is a 'str'." in str(excinfo.value)
-    # 2: optional position argument is invalid
-    with pytest.raises(GenerationError) as excinfo:
-        node.is_valid_location(node, position="invalid_node_argument")
-    assert "The position argument in the psyGen" in str(excinfo.value)
-    assert "method must be one of" in str(excinfo.value)
-    # 3: parents of node and new_node are not the same
-    with pytest.raises(GenerationError) as excinfo:
-        node.is_valid_location(schedule.children[3].children[0])
-    assert ("the node and the location do not have the same "
-            "parent") in str(excinfo.value)
-    # 4: positions are the same
-    prev_node = schedule.children[0]
-    node = schedule.children[1]
-    next_node = schedule.children[2]
-    # a) before this node
-    with pytest.raises(GenerationError) as excinfo:
-        node.is_valid_location(node, position="before")
-    assert "the node and the location are the same" in str(excinfo.value)
-    # b) after this node
-    with pytest.raises(GenerationError) as excinfo:
-        node.is_valid_location(node, position="after")
-    assert "the node and the location are the same" in str(excinfo.value)
-    # c) after previous node
-    with pytest.raises(GenerationError) as excinfo:
-        node.is_valid_location(prev_node, position="after")
-    assert "the node and the location are the same" in str(excinfo.value)
-    # d) before next node
-    with pytest.raises(GenerationError) as excinfo:
-        node.is_valid_location(next_node, position="before")
-    assert "the node and the location are the same" in str(excinfo.value)
-    # 5: valid no previous dependency
-    _, invoke_info = parse(
-        os.path.join(BASE_PATH, "15.14.1_multi_aX_plus_Y_builtin.f90"),
-        api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3", distributed_memory=True).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    schedule = invoke.schedule
-    # 6: valid no prev dep
-    node = schedule.children[2]
-    assert node.is_valid_location(schedule.children[0])
-    # 7: valid prev dep (after)
-    node = schedule.children[6]
-    assert node.is_valid_location(schedule.children[3], position="after")
-    # 8: invalid prev dep (before)
-    assert not node.is_valid_location(schedule.children[3], position="before")
-    # 9: valid no following dep
-    node = schedule.children[4]
-    assert node.is_valid_location(schedule.children[6], position="after")
-    # 10: valid following dep (before)
-    node = schedule.children[0]
-    assert node.is_valid_location(schedule.children[3], position="before")
-    # 11: invalid following dep (after)
-    node = schedule.children[0]
-    assert not node.is_valid_location(schedule.children[3], position="after")
-
-
-def test_node_ancestor():
-    ''' Test the Node.ancestor() method. '''
-    _, invoke = get_invoke("single_invoke.f90", "gocean1.0", idx=0)
-    sched = invoke.schedule
-    kern = sched[0].loop_body[0].loop_body[0]
-    node = kern.ancestor(Node)
-    assert isinstance(node, Schedule)
-    # If 'excluding' is supplied then it can only be a single type or a
-    # tuple of types
-    node = kern.ancestor(Node, excluding=Schedule)
-    assert node is sched[0].loop_body[0]
-    node = kern.ancestor(Node, excluding=(Schedule,))
-    assert node is sched[0].loop_body[0]
-    with pytest.raises(TypeError) as err:
-        kern.ancestor(Node, excluding=[Schedule])
-    assert ("argument to ancestor() must be a type or a tuple of types but "
-            "got: 'list'" in str(err.value))
-    # Check that the include_self argument behaves as expected
-    node = kern.ancestor(Kern, excluding=(Schedule,), include_self=True)
-    assert node is kern
-
-
-def test_dag_names():
-    '''test that the dag_name method returns the correct value for the
-    node class and its specialisations'''
-    _, invoke_info = parse(
-        os.path.join(BASE_PATH, "1_single_invoke.f90"),
-        api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3", distributed_memory=True).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    schedule = invoke.schedule
-    assert super(Schedule, schedule).dag_name == "node_0"
-    assert schedule.dag_name == "schedule_0"
-    assert schedule.children[0].dag_name == "checkHaloExchange(f2)_0"
-    assert schedule.children[3].dag_name == "loop_4"
-    schedule.children[3].loop_type = "colour"
-    assert schedule.children[3].dag_name == "loop_[colour]_4"
-    schedule.children[3].loop_type = ""
-    assert (schedule.children[3].loop_body[0].dag_name ==
-            "kernel_testkern_code_9")
-    _, invoke_info = parse(
-        os.path.join(BASE_PATH, "15.14.3_sum_setval_field_builtin.f90"),
-        api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3", distributed_memory=True).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    schedule = invoke.schedule
-    global_sum = schedule.children[2]
-    assert global_sum.dag_name == "globalsum(asum)_2"
-    builtin = schedule.children[1].loop_body[0]
-    assert builtin.dag_name == "builtin_sum_x_12"
 
 
 def test_openmp_pdo_dag_name():
@@ -2437,93 +1933,6 @@ def test_accenterdatadirective_gencode_4():
         "f3_proxy,f3_proxy%data)\n" in code)
 
 # Class ACCEnterDataDirective end
-
-
-def test_node_dag_no_graphviz(tmpdir, monkeypatch):
-    '''test that dag generation does nothing if graphviz is not
-    installed. We monkeypatch sys.modules to ensure that it always
-    appears that graphviz is not installed on this system. '''
-    import sys
-    monkeypatch.setitem(sys.modules, 'graphviz', None)
-    _, invoke_info = parse(
-        os.path.join(BASE_PATH, "1_single_invoke.f90"),
-        api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3",
-                     distributed_memory=False).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    schedule = invoke.schedule
-    my_file = tmpdir.join('test')
-    schedule.dag(file_name=my_file.strpath)
-    assert not os.path.exists(my_file.strpath)
-
-
-# Use a regex to allow for whitespace differences between graphviz
-# versions. Need a raw-string (r"") to get new-lines handled nicely.
-EXPECTED2 = re.compile(
-    r"digraph {\n"
-    r"\s*schedule_0_start\n"
-    r"\s*schedule_0_end\n"
-    r"\s*loop_1_start\n"
-    r"\s*loop_1_end\n"
-    r"\s*loop_1_end -> loop_7_start \[color=green\]\n"
-    r"\s*schedule_0_start -> loop_1_start \[color=blue\]\n"
-    r"\s*schedule_5_start\n"
-    r"\s*schedule_5_end\n"
-    r"\s*schedule_5_end -> loop_1_end \[color=blue\]\n"
-    r"\s*loop_1_start -> schedule_5_start \[color=blue\]\n"
-    r"\s*kernel_testkern_qr_code_6\n"
-    r"\s*kernel_testkern_qr_code_6 -> schedule_5_end \[color=blue\]\n"
-    r"\s*schedule_5_start -> kernel_testkern_qr_code_6 \[color=blue\]\n"
-    r"\s*loop_7_start\n"
-    r"\s*loop_7_end\n"
-    r"\s*loop_7_end -> schedule_0_end \[color=blue\]\n"
-    r"\s*loop_1_end -> loop_7_start \[color=red\]\n"
-    r"\s*schedule_11_start\n"
-    r"\s*schedule_11_end\n"
-    r"\s*schedule_11_end -> loop_7_end \[color=blue\]\n"
-    r"\s*loop_7_start -> schedule_11_start \[color=blue\]\n"
-    r"\s*kernel_testkern_qr_code_12\n"
-    r"\s*kernel_testkern_qr_code_12 -> schedule_11_end \[color=blue\]\n"
-    r"\s*schedule_11_start -> kernel_testkern_qr_code_12 \[color=blue\]\n"
-    r"}")
-# pylint: enable=anomalous-backslash-in-string
-
-
-def test_node_dag(tmpdir, have_graphviz):
-    '''test that dag generation works correctly. Skip the test if
-    graphviz is not installed'''
-    if not have_graphviz:
-        return
-    _, invoke_info = parse(
-        os.path.join(BASE_PATH, "4.1_multikernel_invokes.f90"),
-        api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3",
-                     distributed_memory=False).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    schedule = invoke.schedule
-    my_file = tmpdir.join('test')
-    schedule.dag(file_name=my_file.strpath)
-    result = my_file.read()
-    assert EXPECTED2.match(result)
-    my_file = tmpdir.join('test.svg')
-    result = my_file.read()
-    for name in ["<title>schedule_0_start</title>",
-                 "<title>schedule_0_end</title>",
-                 "<title>loop_1_start</title>",
-                 "<title>loop_1_end</title>",
-                 "<title>kernel_testkern_qr_code_6</title>",
-                 "<title>kernel_testkern_qr_code_12</title>",
-                 "<svg", "</svg>", ]:
-        assert name in result
-    for colour_name, colour_code in [("blue", "#0000ff"),
-                                     ("green", "#00ff00"),
-                                     ("red", "#ff0000")]:
-        assert colour_name in result or colour_code in result
-
-    with pytest.raises(GenerationError) as excinfo:
-        schedule.dag(file_name=my_file.strpath, file_format="rubbish")
-    assert "unsupported graphviz file format" in str(excinfo.value)
-
 
 def test_haloexchange_halo_depth_get_set():
     '''test that the halo_exchange getter and setter work correctly '''
@@ -2876,46 +2285,6 @@ def test_find_w_args_multiple_deps(monkeypatch, annexed):
     assert len(indices) == vector_size
 
 
-def test_node_abstract_methods():
-    ''' Tests that the abstract methods of the Node class raise appropriate
-    errors. '''
-    _, invoke = get_invoke("single_invoke.f90", "gocean1.0", idx=0)
-    sched = invoke.schedule
-    loop = sched.children[0].loop_body[0]
-    with pytest.raises(NotImplementedError) as err:
-        Node.gen_code(loop, parent=None)
-    assert "Please implement me" in str(err.value)
-
-
-def test_node_coloured_name():
-    ''' Tests for the coloured_name method of the Node class. '''
-    from psyclone.psyir.nodes.node import colored, SCHEDULE_COLOUR_MAP
-    tnode = Node()
-    assert tnode.coloured_name(False) == "Node"
-    # Check that we can change the name of the Node and the colour associated
-    # with it
-    tnode._text_name = "ATest"
-    tnode._colour_key = "Schedule"
-    assert tnode.coloured_name(False) == "ATest"
-    assert tnode.coloured_name(True) == colored(
-        "ATest", SCHEDULE_COLOUR_MAP["Schedule"])
-    # Check that an unrecognised colour-map entry gives us un-coloured text
-    tnode._colour_key = "not-recognised"
-    assert tnode.coloured_name(True) == "ATest"
-
-
-def test_node_str():
-    ''' Tests for the Node.node_str method. '''
-    from psyclone.psyir.nodes.node import colored, SCHEDULE_COLOUR_MAP
-    tnode = Node()
-    # Manually set the colour key for this node to something that will result
-    # in coloured output (if requested *and* termcolor is installed).
-    tnode._colour_key = "Loop"
-    assert tnode.node_str(False) == "Node[]"
-    assert tnode.node_str(True) == colored("Node",
-                                           SCHEDULE_COLOUR_MAP["Loop"]) + "[]"
-
-
 def test_kern_ast():
     ''' Test that we can obtain the fparser2 AST of a kernel. '''
     from psyclone.gocean1p0 import GOKern
@@ -3014,988 +2383,6 @@ def test_dataaccess_same_vector_indices(monkeypatch):
     assert (
         "The halo exchange vector indices for 'd' are the same. This should "
         "never happen" in str(excinfo.value))
-
-
-# Test CodeBlock class
-
-
-def test_codeblock_node_str():
-    ''' Check the node_str method of the Code Block class.'''
-    from psyclone.psyir.nodes.node import colored, SCHEDULE_COLOUR_MAP
-    cblock = CodeBlock([], "dummy")
-    coloredtext = colored("CodeBlock", SCHEDULE_COLOUR_MAP["CodeBlock"])
-    output = cblock.node_str()
-    assert coloredtext+"[" in output
-    assert "]" in output
-
-
-def test_codeblock_can_be_printed():
-    '''Test that a CodeBlock instance can always be printed (i.e. is
-    initialised fully)'''
-    cblock = CodeBlock([], "dummy")
-    assert "CodeBlock[" in str(cblock)
-    assert "]" in str(cblock)
-
-
-def test_codeblock_getastnodes():
-    '''Test that the get_ast_nodes method of a CodeBlock instance returns
-    a copy of the list of nodes from the original AST that are associated with
-    this code block.
-
-    For simplicity we use a list of strings rather than an AST.
-
-    '''
-    original = ["hello", "there"]
-    cblock = CodeBlock(original, CodeBlock.Structure.EXPRESSION)
-    result = cblock.get_ast_nodes
-    assert result == original
-    # Check that the list is a copy not a reference.
-    assert result is not original
-
-
-@pytest.mark.parametrize("structure", [CodeBlock.Structure.STATEMENT,
-                                       CodeBlock.Structure.EXPRESSION])
-def test_codeblock_structure(structure):
-    '''Check that the structure property in the CodeBlock class is set to
-    the provided value.
-
-    '''
-    cblock = CodeBlock([], structure)
-    assert cblock.structure == structure
-
-# Test Loop class
-
-
-def test_loop_navigation_properties():
-    # pylint: disable=too-many-statements
-    ''' Tests the start_expr, stop_expr, step_expr and loop_body
-    setter and getter properties.
-
-    '''
-    loop = Loop()
-
-    # Properties return an error if the node is incomplete
-    error_str = ("Loop malformed or incomplete. It should have exactly 4 "
-                 "children, but found")
-    with pytest.raises(InternalError) as err:
-        _ = loop.start_expr
-    assert error_str in str(err.value)
-
-    # Expressions that are not PSyIR are not accepted
-    with pytest.raises(TypeError) as err:
-        loop.start_expr = "start"
-    assert "Only PSyIR nodes can be assigned as the Loop start expression" \
-        ", but found '" in str(err.value)
-    with pytest.raises(TypeError) as err:
-        loop.stop_expr = "stop"
-    assert "Only PSyIR nodes can be assigned as the Loop stop expression" \
-        ", but found '" in str(err.value)
-    with pytest.raises(TypeError) as err:
-        loop.step_expr = "step"
-    assert "Only PSyIR nodes can be assigned as the Loop step expression" \
-        ", but found '" in str(err.value)
-
-    loop.addchild(Literal("start", DataType.INTEGER, parent=loop))
-    loop.addchild(Literal("stop", DataType.INTEGER, parent=loop))
-    loop.addchild(Literal("step", DataType.INTEGER, parent=loop))
-
-    # If it's not fully complete, it still returns an error
-    with pytest.raises(InternalError) as err:
-        _ = loop.start_expr
-    assert error_str in str(err.value)
-    with pytest.raises(InternalError) as err:
-        _ = loop.stop_expr
-    assert error_str in str(err.value)
-    with pytest.raises(InternalError) as err:
-        _ = loop.step_expr
-    assert error_str in str(err.value)
-    with pytest.raises(InternalError) as err:
-        _ = loop.loop_body
-    assert error_str in str(err.value)
-    with pytest.raises(InternalError) as err:
-        loop.start_expr = Literal("invalid", DataType.INTEGER, parent=loop)
-    assert error_str in str(err.value)
-    with pytest.raises(InternalError) as err:
-        loop.stop_expr = Literal("invalid", DataType.INTEGER, parent=loop)
-    assert error_str in str(err.value)
-    with pytest.raises(InternalError) as err:
-        loop.step_expr = Literal("invalid", DataType.INTEGER, parent=loop)
-    assert error_str in str(err.value)
-
-    # The fourth child has to be a Schedule
-    loop.addchild(Literal("loop_body", DataType.INTEGER, parent=loop))
-    with pytest.raises(InternalError) as err:
-        _ = loop.loop_body
-    assert "Loop malformed or incomplete. Fourth child should be a " \
-        "Schedule node, but found loop with " in str(err.value)
-
-    # Fix loop and check that Getters properties work
-    del loop.children[3]
-    loop.addchild(Schedule(parent=loop))
-    loop.loop_body.addchild(Return(parent=loop.loop_body))
-
-    assert loop.start_expr.value == "start"
-    assert loop.stop_expr.value == "stop"
-    assert loop.step_expr.value == "step"
-    assert isinstance(loop.loop_body[0], Return)
-
-    # Test Setters
-    loop.start_expr = Literal("newstart", DataType.INTEGER, parent=loop)
-    loop.stop_expr = Literal("newstop", DataType.INTEGER, parent=loop)
-    loop.step_expr = Literal("newstep", DataType.INTEGER, parent=loop)
-
-    assert loop.start_expr.value == "newstart"
-    assert loop.stop_expr.value == "newstop"
-    assert loop.step_expr.value == "newstep"
-
-
-def test_loop_invalid_type():
-    ''' Tests assigning an invalid type to a Loop object. '''
-    _, invoke = get_invoke("single_invoke.f90", "gocean1.0", idx=0)
-    sched = invoke.schedule
-    loop = sched.children[0].loop_body[0]
-    assert isinstance(loop, Loop)
-    with pytest.raises(GenerationError) as err:
-        loop.loop_type = "not_a_valid_type"
-    assert ("loop_type value (not_a_valid_type) is invalid. Must be one of "
-            "['inner', 'outer']" in str(err.value))
-
-
-def test_loop_gen_code():
-    ''' Check that the Loop gen_code method prints the proper loop '''
-    _, invoke_info = parse(os.path.join(BASE_PATH,
-                                        "1.0.1_single_named_invoke.f90"),
-                           api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3", distributed_memory=True).create(invoke_info)
-
-    # By default DynLoop has step = 1 and it is not printed in the Fortran DO
-    gen = str(psy.gen)
-    assert "DO cell=1,mesh%get_last_halo_cell(1)" in gen
-
-    # Change step to 2
-    loop = psy.invokes.get('invoke_important_invoke').schedule[3]
-    loop.step_expr = Literal("2", DataType.INTEGER, parent=loop)
-
-    # Now it is printed in the Fortran DO with the expression  ",2" at the end
-    gen = str(psy.gen)
-    assert "DO cell=1,mesh%get_last_halo_cell(1),2" in gen
-
-
-def test_invalid_loop_annotations():
-    ''' Check that the Loop constructor validates any supplied annotations. '''
-    # Check that we can have 'was_where' on its own
-    test_loop = Loop(annotations=['was_where'])
-    assert test_loop.annotations == ['was_where']
-    # Check that 'was_single_stmt' on its own raises an error
-    with pytest.raises(InternalError) as err:
-        Loop(annotations=['was_single_stmt'])
-    assert ("Loop with the 'was_single_stmt' annotation must also have the "
-            "'was_where'" in str(err.value))
-    # Check that it's accepted in combination with 'was_where'
-    test_loop = Loop(annotations=['was_single_stmt', 'was_where'])
-    assert test_loop.annotations == ['was_single_stmt', 'was_where']
-
-
-# Test IfBlock class
-
-def test_ifblock_invalid_annotation():
-    ''' Test that initialising IfBlock with invalid annotations produce the
-    expected error.'''
-
-    with pytest.raises(InternalError) as err:
-        _ = IfBlock(annotations=["invalid"])
-    assert ("IfBlock with unrecognized annotation 'invalid', valid "
-            "annotations are:") in str(err.value)
-
-
-def test_ifblock_node_str():
-    ''' Check the node_str method of the IfBlock class.'''
-    from psyclone.psyir.nodes.node import colored, SCHEDULE_COLOUR_MAP
-    colouredif = colored("If", SCHEDULE_COLOUR_MAP["If"])
-
-    ifblock = IfBlock()
-    output = ifblock.node_str()
-    assert colouredif+"[]" in output
-
-    ifblock = IfBlock(annotations=['was_elseif'])
-    output = ifblock.node_str()
-    assert colouredif+"[annotations='was_elseif']" in output
-
-
-def test_ifblock_view_indices(capsys):
-    ''' Check that the view method only displays indices on the nodes
-    in the body (and else body) of an IfBlock. '''
-    from psyclone.psyir.nodes.node import colored, SCHEDULE_COLOUR_MAP
-    colouredif = colored("If", SCHEDULE_COLOUR_MAP["If"])
-    colouredreturn = colored("Return", SCHEDULE_COLOUR_MAP["Return"])
-    colouredref = colored("Reference", SCHEDULE_COLOUR_MAP["Reference"])
-
-    ifblock = IfBlock()
-    ref1 = Reference('condition1', parent=ifblock)
-    ifblock.addchild(ref1)
-    sch = Schedule(parent=ifblock)
-    ifblock.addchild(sch)
-    ret = Return(parent=sch)
-    sch.addchild(ret)
-    ifblock.view()
-    output, _ = capsys.readouterr()
-    # Check that we only prepend child indices where it makes sense
-    assert colouredif + "[]" in output
-    assert "0: " + colouredreturn in output
-    assert ": " + colouredref not in output
-
-
-def test_ifblock_can_be_printed():
-    '''Test that an IfBlock instance can always be printed (i.e. is
-    initialised fully)'''
-    ifblock = IfBlock()
-    ref1 = Reference('condition1', parent=ifblock)
-    ifblock.addchild(ref1)
-    sch = Schedule(parent=ifblock)
-    ifblock.addchild(sch)
-    ret = Return(parent=sch)
-    sch.addchild(ret)
-
-    assert "If[]\n" in str(ifblock)
-    assert "condition1" in str(ifblock)  # Test condition is printed
-    assert "Return[]" in str(ifblock)  # Test if_body is printed
-
-
-def test_ifblock_properties():
-    '''Test that an IfBlock node properties can be retrieved'''
-    ifblock = IfBlock()
-
-    # Condition can't be retrieved before is added as a child.
-    with pytest.raises(InternalError) as err:
-        _ = ifblock.condition
-    assert("IfBlock malformed or incomplete. It should have "
-           "at least 2 children, but found 0." in str(err.value))
-
-    ref1 = Reference('condition1', parent=ifblock)
-    ifblock.addchild(ref1)
-
-    # If_body can't be retrieved before is added as a child.
-    with pytest.raises(InternalError) as err:
-        _ = ifblock.if_body
-    assert("IfBlock malformed or incomplete. It should have "
-           "at least 2 children, but found 1." in str(err.value))
-
-    sch = Schedule(parent=ifblock)
-    ifblock.addchild(sch)
-    ret = Return(parent=sch)
-    sch.addchild(ret)
-
-    # Now we can retrieve the condition and the if_body, but else is empty
-    assert ifblock.condition is ref1
-    assert ifblock.if_body[0] is ret
-    assert not ifblock.else_body
-
-    sch2 = Schedule(parent=ifblock)
-    ifblock.addchild(sch2)
-    ret2 = Return(parent=sch2)
-    sch2.addchild(ret2)
-
-    # Now we can retrieve else_body
-    assert ifblock.else_body[0] is ret2
-
-
-def test_ifblock_create():
-    '''Test that the create method in an IfBlock class correctly creates
-    an IfBlock instance.
-
-    '''
-    # Without an else clause.
-    if_condition = Literal('true', DataType.BOOLEAN)
-    if_body = [Assignment.create(Reference("tmp"),
-                                 Literal("0.0", DataType.REAL)),
-               Assignment.create(Reference("tmp2"),
-                                 Literal("1.0", DataType.REAL))]
-    ifblock = IfBlock.create(if_condition, if_body)
-    if_schedule = ifblock.children[1]
-    assert isinstance(if_schedule, Schedule)
-    check_links(ifblock, [if_condition, if_schedule])
-    check_links(if_schedule, if_body)
-    result = FortranWriter().ifblock_node(ifblock)
-    assert result == ("if (.true.) then\n"
-                      "  tmp=0.0\n"
-                      "  tmp2=1.0\n"
-                      "end if\n")
-
-    # With an else clause.
-    else_body = [Assignment.create(Reference("tmp"),
-                                   Literal("1.0", DataType.REAL)),
-                 Assignment.create(Reference("tmp2"),
-                                   Literal("0.0", DataType.REAL))]
-    ifblock = IfBlock.create(if_condition, if_body, else_body)
-    if_schedule = ifblock.children[1]
-    assert isinstance(if_schedule, Schedule)
-    else_schedule = ifblock.children[2]
-    assert isinstance(else_schedule, Schedule)
-    check_links(ifblock, [if_condition, if_schedule, else_schedule])
-    check_links(if_schedule, if_body)
-    check_links(else_schedule, else_body)
-    result = FortranWriter().ifblock_node(ifblock)
-    assert result == ("if (.true.) then\n"
-                      "  tmp=0.0\n"
-                      "  tmp2=1.0\n"
-                      "else\n"
-                      "  tmp=1.0\n"
-                      "  tmp2=0.0\n"
-                      "end if\n")
-
-
-def test_ifblock_create_invalid():
-    '''Test that the create method in an IfBlock class raises the expected
-    exception if the provided input is invalid.
-
-    '''
-    if_condition = Literal('true', DataType.BOOLEAN)
-    if_body = [Assignment.create(Reference("tmp"),
-                                 Literal("0.0", DataType.REAL)),
-               Assignment.create(Reference("tmp2"),
-                                 Literal("1.0", DataType.REAL))]
-
-    # if_condition not a Node.
-    with pytest.raises(GenerationError) as excinfo:
-        _ = IfBlock.create("True", "invalid")
-    assert ("if_condition argument in create method of IfBlock class should "
-            "be a PSyIR Node but found 'str'.") in str(excinfo.value)
-
-    # One of more if body not a Node.
-    if_body_err = [Assignment.create(Reference("tmp"),
-                                     Literal("0.0", DataType.REAL)),
-                   "invalid"]
-    with pytest.raises(GenerationError) as excinfo:
-        _ = IfBlock.create(if_condition, if_body_err)
-    assert ("if_body argument in create method of IfBlock class should be a "
-            "list of PSyIR Nodes but it is either not a list or one of the "
-            "list's children is not a Node.") in str(excinfo.value)
-
-    # If body not a list.
-    with pytest.raises(GenerationError) as excinfo:
-        _ = IfBlock.create(if_condition, "invalid")
-    assert ("if_body argument in create method of IfBlock class should be a "
-            "list of PSyIR Nodes but it is either not a list or one of the "
-            "list's children is not a Node.") in str(excinfo.value)
-
-    # One of more of else_body not a Node.
-    else_body_err = [Assignment.create(Reference("tmp"),
-                                       Literal("1.0", DataType.REAL)),
-                     "invalid"]
-    with pytest.raises(GenerationError) as excinfo:
-        _ = IfBlock.create(if_condition, if_body, else_body_err)
-    assert ("else_body argument in create method of IfBlock class should be a "
-            "list of PSyIR Nodes but it is either not a list or one of the "
-            "list's children is not a Node.") in str(excinfo.value)
-
-    # Else body not a list.
-    with pytest.raises(GenerationError) as excinfo:
-        _ = IfBlock.create(if_condition, if_body, "invalid")
-    assert ("else_body argument in create method of IfBlock class should be a "
-            "list of PSyIR Nodes but it is either not a list or one of the "
-            "list's children is not a Node.") in str(excinfo.value)
-
-
-# Test Assignment class
-
-def test_assignment_node_str():
-    ''' Check the node_str method of the Assignment class.'''
-    from psyclone.psyir.nodes.node import colored, SCHEDULE_COLOUR_MAP
-
-    assignment = Assignment()
-    coloredtext = colored("Assignment", SCHEDULE_COLOUR_MAP["Assignment"])
-    assert coloredtext+"[]" in assignment.node_str()
-
-
-def test_assignment_can_be_printed():
-    '''Test that an Assignment instance can always be printed (i.e. is
-    initialised fully)'''
-    assignment = Assignment()
-    assert "Assignment[]\n" in str(assignment)
-
-
-def test_assignment_semantic_navigation():
-    '''Test that the Assignment navigation properties reference the expected
-    children'''
-    assignment = Assignment()
-
-    # lhs should fail if first child is not present
-    with pytest.raises(InternalError) as err:
-        _ = assignment.lhs
-    assert "' malformed or incomplete. It needs at least 1 child to have " \
-        "a lhs." in str(err.value)
-
-    ref = Reference("a", assignment)
-    assignment.addchild(ref)
-
-    # rhs should fail if second child is not present
-    with pytest.raises(InternalError) as err:
-        _ = assignment.rhs
-    assert " malformed or incomplete. It needs at least 2 children to have " \
-        "a rhs." in str(err.value)
-
-    lit = Literal("1", DataType.INTEGER, assignment)
-    assignment.addchild(lit)
-    assert assignment.lhs is assignment._children[0]
-    assert assignment.rhs is assignment._children[1]
-
-
-def test_assignment_create():
-    '''Test that the create method in the Assignment class correctly
-    creates an Assignment instance.
-
-    '''
-    lhs = Reference("tmp")
-    rhs = Literal("0.0", DataType.REAL)
-    assignment = Assignment.create(lhs, rhs)
-    check_links(assignment, [lhs, rhs])
-    result = FortranWriter().assignment_node(assignment)
-    assert result == "tmp=0.0\n"
-
-
-def test_assignment_create_invalid():
-    '''Test that the create method in an Assignment class raises the expected
-    exception if the provided input is invalid.
-
-    '''
-    # lhs not a Node.
-    with pytest.raises(GenerationError) as excinfo:
-        _ = Assignment.create("invalid", Literal("0.0", DataType.REAL))
-    assert ("lhs argument in create method of Assignment class should "
-            "be a PSyIR Node but found 'str'.") in str(excinfo.value)
-
-    # rhs not a Node.
-    with pytest.raises(GenerationError) as excinfo:
-        _ = Assignment.create(Reference("tmp"), "invalid")
-    assert ("rhs argument in create method of Assignment class should "
-            "be a PSyIR Node but found 'str'.") in str(excinfo.value)
-
-
-# Test Reference class
-
-def test_reference_node_str():
-    ''' Check the node_str method of the Reference class.'''
-    from psyclone.psyir.nodes.node import colored, SCHEDULE_COLOUR_MAP
-    kschedule = KernelSchedule("kname")
-    kschedule.symbol_table.add(DataSymbol("rname", DataType.INTEGER))
-    assignment = Assignment(parent=kschedule)
-    ref = Reference("rname", assignment)
-    coloredtext = colored("Reference", SCHEDULE_COLOUR_MAP["Reference"])
-    assert coloredtext+"[name:'rname']" in ref.node_str()
-
-
-def test_reference_can_be_printed():
-    '''Test that a Reference instance can always be printed (i.e. is
-    initialised fully)'''
-    kschedule = KernelSchedule("kname")
-    kschedule.symbol_table.add(DataSymbol("rname", DataType.INTEGER))
-    assignment = Assignment(parent=kschedule)
-    ref = Reference("rname", assignment)
-    assert "Reference[name:'rname']" in str(ref)
-
-
-def test_reference_optional_parent():
-    '''Test that the parent attribute is None if the optional parent
-    argument is not supplied.
-
-    '''
-    ref = Reference("rname")
-    assert ref.parent is None
-
-
-def test_reference_symbol(monkeypatch):
-    '''Test that the symbol method in a Reference Node instance returns
-    the associated symbol if there is one and None if not. Also test
-    for an incorrect scope argument.
-
-    '''
-    _, invoke = get_invoke("single_invoke_kern_with_global.f90",
-                           api="gocean1.0", idx=0)
-    sched = invoke.schedule
-    kernels = sched.walk(Kern)
-    kernel_schedule = kernels[0].get_kernel_schedule()
-    references = kernel_schedule.walk(Reference)
-
-    # Symbol in KernelSchedule SymbolTable
-    field_old = references[0]
-    assert field_old.name == "field_old"
-    assert isinstance(field_old.symbol(), DataSymbol)
-    assert field_old.symbol().name == field_old.name
-
-    # Symbol in KernelSchedule SymbolTable with KernelSchedule scope
-    assert isinstance(field_old.symbol(scope_limit=kernel_schedule),
-                      DataSymbol)
-    assert field_old.symbol().name == field_old.name
-
-    # Symbol in KernelSchedule SymbolTable with parent scope
-    assert field_old.symbol(scope_limit=field_old.parent) is None
-
-    # Symbol in Container SymbolTable
-    alpha = references[6]
-    assert alpha.name == "alpha"
-    assert isinstance(alpha.symbol(), DataSymbol)
-    assert alpha.symbol().name == alpha.name
-
-    # Symbol in Container SymbolTable with KernelSchedule scope
-    assert alpha.symbol(scope_limit=kernel_schedule) is None
-
-    # Symbol in Container SymbolTable with Container scope
-    assert isinstance(kernel_schedule.root, Container)
-    assert alpha.symbol(scope_limit=kernel_schedule.root).name == alpha.name
-
-    # Symbol method with invalid scope type
-    with pytest.raises(TypeError) as excinfo:
-        _ = alpha.symbol(scope_limit="hello")
-    assert ("The scope_limit argument 'hello' provided to the symbol method, "
-            "is not of type `Node`." in str(excinfo.value))
-
-    # Symbol method with invalid scope location
-    with pytest.raises(ValueError) as excinfo:
-        _ = alpha.symbol(scope_limit=alpha)
-    assert ("The scope_limit node 'Reference[name:'alpha']' provided to the "
-            "symbol method, is not an ancestor of this reference node "
-            "'Reference[name:'alpha']'." in str(excinfo.value))
-
-    # Symbol not in any container (rename alpha to something that is
-    # not defined)
-    monkeypatch.setattr(alpha, "_reference", "not_defined")
-    assert not alpha.symbol()
-
-# Test Array class
-
-
-def test_array_node_str():
-    ''' Check the node_str method of the Array class.'''
-    from psyclone.psyir.nodes.node import colored, SCHEDULE_COLOUR_MAP
-    kschedule = KernelSchedule("kname")
-    kschedule.symbol_table.add(DataSymbol("aname", DataType.INTEGER,
-                                          [DataSymbol.Extent.ATTRIBUTE]))
-    assignment = Assignment(parent=kschedule)
-    array = Array("aname", parent=assignment)
-    coloredtext = colored("ArrayReference", SCHEDULE_COLOUR_MAP["Reference"])
-    assert coloredtext+"[name:'aname']" in array.node_str()
-
-
-def test_array_can_be_printed():
-    '''Test that an Array instance can always be printed (i.e. is
-    initialised fully)'''
-    kschedule = KernelSchedule("kname")
-    kschedule.symbol_table.add(DataSymbol("aname", DataType.INTEGER))
-    assignment = Assignment(parent=kschedule)
-    array = Array("aname", assignment)
-    assert "ArrayReference[name:'aname']\n" in str(array)
-
-
-def test_array_create():
-    '''Test that the create method in the Array class correctly
-    creates an Array instance.
-
-    '''
-    children = [Reference("i"), Reference("j"), Literal("1", DataType.REAL)]
-    array = Array.create("temp", children)
-    check_links(array, children)
-    result = FortranWriter().array_node(array)
-    assert result == "temp(i,j,1)"
-
-
-def test_array_create_invalid():
-    '''Test that the create method in an Array class raises the expected
-    exception if the provided input is invalid.
-
-    '''
-    # name is not a string
-    with pytest.raises(GenerationError) as excinfo:
-        _ = Array.create([], [])
-    assert ("name argument in create method of Array class should "
-            "be a string but found 'list'."
-            in str(excinfo.value))
-
-    # name is an empty string
-    with pytest.raises(GenerationError) as excinfo:
-        _ = Array.create("", [])
-    assert ("name argument in create method of Array class can't "
-            "be an empty string.")
-
-    # children not a list
-    with pytest.raises(GenerationError) as excinfo:
-        _ = Array.create("temp", "invalid")
-    assert ("children argument in create method of Array class should "
-            "be a list but found 'str'." in str(excinfo.value))
-
-    # contents of children list are not Node
-    with pytest.raises(GenerationError) as excinfo:
-        _ = Array.create("temp",
-                         [Reference("i"), "invalid"])
-    assert (
-        "child of children argument in create method of Array class "
-        "should be a PSyIR Node but found 'str'." in str(excinfo.value))
-
-
-# Test BinaryOperation class
-def test_binaryoperation_initialization():
-    ''' Check the initialization method of the BinaryOperation class works
-    as expected.'''
-
-    with pytest.raises(TypeError) as err:
-        _ = BinaryOperation("not an operator")
-    assert "BinaryOperation operator argument must be of type " \
-           "BinaryOperation.Operator but found" in str(err.value)
-    bop = BinaryOperation(BinaryOperation.Operator.ADD)
-    assert bop._operator is BinaryOperation.Operator.ADD
-
-
-def test_binaryoperation_operator():
-    '''Test that the operator property returns the binaryoperator in the
-    binaryoperation.
-
-    '''
-    binary_operation = BinaryOperation(BinaryOperation.Operator.ADD)
-    assert binary_operation.operator == BinaryOperation.Operator.ADD
-
-
-def test_binaryoperation_node_str():
-    ''' Check the node_str method of the Binary Operation class.'''
-    from psyclone.psyir.nodes.node import colored, SCHEDULE_COLOUR_MAP
-    binary_operation = BinaryOperation(BinaryOperation.Operator.ADD)
-    op1 = Literal("1", DataType.INTEGER, parent=binary_operation)
-    op2 = Literal("1", DataType.INTEGER, parent=binary_operation)
-    binary_operation.addchild(op1)
-    binary_operation.addchild(op2)
-    coloredtext = colored("BinaryOperation",
-                          SCHEDULE_COLOUR_MAP["Operation"])
-    assert coloredtext+"[operator:'ADD']" in binary_operation.node_str()
-
-
-def test_binaryoperation_can_be_printed():
-    '''Test that a Binary Operation instance can always be printed (i.e. is
-    initialised fully)'''
-    binary_operation = BinaryOperation(BinaryOperation.Operator.ADD)
-    assert "BinaryOperation[operator:'ADD']" in str(binary_operation)
-    op1 = Literal("1", DataType.INTEGER, parent=binary_operation)
-    op2 = Literal("2", DataType.INTEGER, parent=binary_operation)
-    binary_operation.addchild(op1)
-    binary_operation.addchild(op2)
-    # Check the node children are also printed
-    assert "Literal[value:'1', DataType.INTEGER]\n" in str(binary_operation)
-    assert "Literal[value:'2', DataType.INTEGER]" in str(binary_operation)
-
-
-def test_binaryoperation_create():
-    '''Test that the create method in the BinaryOperation class correctly
-    creates a BinaryOperation instance.
-
-    '''
-    lhs = Reference("tmp1")
-    rhs = Reference("tmp2")
-    oper = BinaryOperation.Operator.ADD
-    binaryoperation = BinaryOperation.create(oper, lhs, rhs)
-    check_links(binaryoperation, [lhs, rhs])
-    result = FortranWriter().binaryoperation_node(binaryoperation)
-    assert result == "tmp1 + tmp2"
-
-
-def test_binaryoperation_create_invalid():
-    '''Test that the create method in a BinaryOperation class raises the
-    expected exception if the provided input is invalid.
-
-    '''
-    ref1 = Reference("tmp1")
-    ref2 = Reference("tmp2")
-    add = BinaryOperation.Operator.ADD
-
-    # oper not a BinaryOperation.Operator.
-    with pytest.raises(GenerationError) as excinfo:
-        _ = BinaryOperation.create("invalid", ref1, ref2)
-    assert ("oper argument in create method of BinaryOperation class should "
-            "be a PSyIR BinaryOperation Operator but found 'str'."
-            in str(excinfo.value))
-
-    # lhs not a Node.
-    with pytest.raises(GenerationError) as excinfo:
-        _ = BinaryOperation.create(add, "invalid", ref2)
-    assert ("lhs argument in create method of BinaryOperation class should "
-            "be a PSyIR Node but found 'str'.") in str(excinfo.value)
-
-    # rhs not a Node.
-    with pytest.raises(GenerationError) as excinfo:
-        _ = BinaryOperation.create(add, ref1, "invalid")
-    assert ("rhs argument in create method of BinaryOperation class should "
-            "be a PSyIR Node but found 'str'.") in str(excinfo.value)
-
-
-# Test UnaryOperation class
-def test_unaryoperation_initialization():
-    ''' Check the initialization method of the UnaryOperation class works
-    as expected.'''
-
-    with pytest.raises(TypeError) as err:
-        _ = UnaryOperation("not an operator")
-    assert "UnaryOperation operator argument must be of type " \
-           "UnaryOperation.Operator but found" in str(err.value)
-    uop = UnaryOperation(UnaryOperation.Operator.MINUS)
-    assert uop._operator is UnaryOperation.Operator.MINUS
-
-
-def test_unaryoperation_operator():
-    '''Test that the operator property returns the unaryoperator in the
-    unaryoperation.
-
-    '''
-    unary_operation = UnaryOperation(UnaryOperation.Operator.MINUS)
-    assert unary_operation.operator == UnaryOperation.Operator.MINUS
-
-
-def test_unaryoperation_node_str():
-    ''' Check the view method of the UnaryOperation class.'''
-    from psyclone.psyir.nodes.node import colored, SCHEDULE_COLOUR_MAP
-    unary_operation = UnaryOperation(UnaryOperation.Operator.MINUS)
-    ref1 = Reference("a", parent=unary_operation)
-    unary_operation.addchild(ref1)
-    coloredtext = colored("UnaryOperation",
-                          SCHEDULE_COLOUR_MAP["Operation"])
-    assert coloredtext+"[operator:'MINUS']" in unary_operation.node_str()
-
-
-def test_unaryoperation_can_be_printed():
-    '''Test that a UnaryOperation instance can always be printed (i.e. is
-    initialised fully)'''
-    unary_operation = UnaryOperation(UnaryOperation.Operator.MINUS)
-    assert "UnaryOperation[operator:'MINUS']" in str(unary_operation)
-    op1 = Literal("1", DataType.INTEGER, parent=unary_operation)
-    unary_operation.addchild(op1)
-    # Check the node children are also printed
-    assert "Literal[value:'1', DataType.INTEGER]" in str(unary_operation)
-
-
-def test_unaryoperation_create():
-    '''Test that the create method in the UnaryOperation class correctly
-    creates a UnaryOperation instance.
-
-    '''
-    child = Reference("tmp")
-    oper = UnaryOperation.Operator.SIN
-    unaryoperation = UnaryOperation.create(oper, child)
-    check_links(unaryoperation, [child])
-    result = FortranWriter().unaryoperation_node(unaryoperation)
-    assert result == "SIN(tmp)"
-
-
-def test_unaryoperation_create_invalid():
-    '''Test that the create method in a UnaryOperation class raises the
-    expected exception if the provided input is invalid.
-
-    '''
-    # oper not a UnaryOperator.Operator.
-    with pytest.raises(GenerationError) as excinfo:
-        _ = UnaryOperation.create("invalid", Reference("tmp"))
-    assert ("oper argument in create method of UnaryOperation class should "
-            "be a PSyIR UnaryOperation Operator but found 'str'."
-            in str(excinfo.value))
-
-    # child not a Node.
-    with pytest.raises(GenerationError) as excinfo:
-        _ = UnaryOperation.create(UnaryOperation.Operator.SIN, "invalid")
-    assert ("child argument in create method of UnaryOperation class should "
-            "be a PSyIR Node but found 'str'.") in str(excinfo.value)
-
-
-def test_naryoperation_node_str():
-    ''' Check the node_str method of the Nary Operation class.'''
-    from psyclone.psyir.nodes.node import colored, SCHEDULE_COLOUR_MAP
-    nary_operation = NaryOperation(NaryOperation.Operator.MAX)
-    nary_operation.addchild(Literal("1", DataType.INTEGER,
-                                    parent=nary_operation))
-    nary_operation.addchild(Literal("1", DataType.INTEGER,
-                                    parent=nary_operation))
-    nary_operation.addchild(Literal("1", DataType.INTEGER,
-                                    parent=nary_operation))
-
-    coloredtext = colored("NaryOperation",
-                          SCHEDULE_COLOUR_MAP["Operation"])
-    assert coloredtext+"[operator:'MAX']" in nary_operation.node_str()
-
-
-def test_naryoperation_can_be_printed():
-    '''Test that an Nary Operation instance can always be printed (i.e. is
-    initialised fully)'''
-    nary_operation = NaryOperation(NaryOperation.Operator.MAX)
-    assert "NaryOperation[operator:'MAX']" in str(nary_operation)
-    nary_operation.addchild(Literal("1", DataType.INTEGER,
-                                    parent=nary_operation))
-    nary_operation.addchild(Literal("2", DataType.INTEGER,
-                                    parent=nary_operation))
-    nary_operation.addchild(Literal("3", DataType.INTEGER,
-                                    parent=nary_operation))
-    # Check the node children are also printed
-    assert "Literal[value:'1', DataType.INTEGER]\n" in str(nary_operation)
-    assert "Literal[value:'2', DataType.INTEGER]\n" in str(nary_operation)
-    assert "Literal[value:'3', DataType.INTEGER]" in str(nary_operation)
-
-
-def test_naryoperation_create():
-    '''Test that the create method in the NaryOperation class correctly
-    creates an NaryOperation instance.
-
-    '''
-    children = [Reference("tmp1"), Reference("tmp2"), Reference("tmp3")]
-    oper = NaryOperation.Operator.MAX
-    naryoperation = NaryOperation.create(oper, children)
-    check_links(naryoperation, children)
-    result = FortranWriter().naryoperation_node(naryoperation)
-    assert result == "MAX(tmp1, tmp2, tmp3)"
-
-
-def test_naryoperation_create_invalid():
-    '''Test that the create method in an NaryOperation class raises the
-    expected exception if the provided input is invalid.
-
-    '''
-    # oper not an NaryOperation.Operator
-    with pytest.raises(GenerationError) as excinfo:
-        _ = NaryOperation.create("invalid", [])
-    assert ("oper argument in create method of NaryOperation class should "
-            "be a PSyIR NaryOperation Operator but found 'str'."
-            in str(excinfo.value))
-
-    # children not a list
-    with pytest.raises(GenerationError) as excinfo:
-        _ = NaryOperation.create(NaryOperation.Operator.SUM, "invalid")
-    assert ("children argument in create method of NaryOperation class should "
-            "be a list but found 'str'." in str(excinfo.value))
-
-    # contents of children list are not Node
-    with pytest.raises(GenerationError) as excinfo:
-        _ = NaryOperation.create(NaryOperation.Operator.SUM,
-                                 [Reference("tmp1"), "invalid"])
-    assert (
-        "child of children argument in create method of NaryOperation class "
-        "should be a PSyIR Node but found 'str'." in str(excinfo.value))
-
-
-# Test Return class
-
-def test_return_node_str():
-    ''' Check the node_str method of the Return class.'''
-    from psyclone.psyir.nodes.node import colored, SCHEDULE_COLOUR_MAP
-    return_stmt = Return()
-    coloredtext = colored("Return", SCHEDULE_COLOUR_MAP["Return"])
-    assert coloredtext+"[]" in return_stmt.node_str()
-
-
-def test_return_can_be_printed():
-    '''Test that a Return instance can always be printed (i.e. is
-    initialised fully)'''
-    return_stmt = Return()
-    assert "Return[]\n" in str(return_stmt)
-
-
-# Test Container class
-
-def test_container_init():
-    '''Test that a container is initialised as expected.'''
-    container = Container("test")
-    assert container._name == "test"
-    assert container._parent is None
-    assert isinstance(container._symbol_table, SymbolTable)
-
-
-def test_container_init_parent():
-    '''Test that a container parent argument is stored as expected.'''
-    container = Container("test", parent="hello")
-    assert container.parent == "hello"
-
-
-def test_container_name():
-    '''Test that the container name can be set and changed as
-    expected.'''
-    container = Container("test")
-    assert container.name == "test"
-    container.name = "new_test"
-    assert container.name == "new_test"
-
-
-def test_container_symbol_table():
-    '''Test that the container symbol_table method returns the expected
-    content.'''
-    container = Container("test")
-    assert isinstance(container._symbol_table, SymbolTable)
-    assert container.symbol_table is container._symbol_table
-
-
-def test_container_node_str():
-    '''Check the node_str method of the Container class.'''
-    from psyclone.psyir.nodes.node import colored, SCHEDULE_COLOUR_MAP
-    cont_stmt = Container("bin")
-    coloredtext = colored("Container", SCHEDULE_COLOUR_MAP["Container"])
-    assert coloredtext+"[bin]" in cont_stmt.node_str()
-
-
-def test_container_can_be_printed():
-    '''Test that a Container instance can always be printed (i.e. is
-    initialised fully)'''
-    cont_stmt = Container("box")
-    assert "Container[box]\n" in str(cont_stmt)
-
-
-def test_container_create():
-    '''Test that the create method in the Container class correctly
-    creates a Container instance.
-
-    '''
-    symbol_table = SymbolTable()
-    symbol_table.add(DataSymbol("tmp", DataType.REAL))
-    kernel1 = KernelSchedule.create("mod_1", SymbolTable(), [])
-    kernel2 = KernelSchedule.create("mod_2", SymbolTable(), [])
-    container = Container.create("container_name", symbol_table,
-                                 [kernel1, kernel2])
-    check_links(container, [kernel1, kernel2])
-    assert container.symbol_table is symbol_table
-    result = FortranWriter().container_node(container)
-    assert result == (
-        "module container_name\n"
-        "  real :: tmp\n\n"
-        "  contains\n"
-        "  subroutine mod_1()\n\n\n"
-        "  end subroutine mod_1\n"
-        "  subroutine mod_2()\n\n\n"
-        "  end subroutine mod_2\n\n"
-        "end module container_name\n")
-
-
-def test_container_create_invalid():
-    '''Test that the create method in a Container class raises the
-    expected exception if the provided input is invalid.
-
-    '''
-    symbol_table = SymbolTable()
-    symbol_table.add(DataSymbol("x", DataType.REAL))
-    children = [KernelSchedule.create("mod_1", SymbolTable(), [])]
-
-    # name is not a string.
-    with pytest.raises(GenerationError) as excinfo:
-        _ = Container.create(1, symbol_table, children)
-    assert ("name argument in create method of Container class "
-            "should be a string but found 'int'.") in str(excinfo.value)
-
-    # symbol_table not a SymbolTable.
-    with pytest.raises(GenerationError) as excinfo:
-        _ = Container.create("container", "invalid", children)
-    assert ("symbol_table argument in create method of Container class "
-            "should be a SymbolTable but found 'str'.") in str(excinfo.value)
-
-    # children not a list.
-    with pytest.raises(GenerationError) as excinfo:
-        _ = Container.create("mod_name", symbol_table, "invalid")
-    assert ("children argument in create method of Container class should "
-            "be a list but found 'str'." in str(excinfo.value))
-
-    # contents of children list are not Container or KernelSchedule.
-    with pytest.raises(GenerationError) as excinfo:
-        _ = Container.create("mod_name", symbol_table, ["invalid"])
-    assert (
-        "child of children argument in create method of Container class "
-        "should be a PSyIR KernelSchedule or Container but found 'str'."
-        in str(excinfo.value))
 
 
 # Test KernelSchedule Class
