+++ src/psyclone/tests/psyir/nodes/schedule_test.py
@@ -1,0 +1,94 @@
+# -----------------------------------------------------------------------------
+# BSD 3-Clause License
+#
+# Copyright (c) 2019-2020, Science and Technology Facilities Council.
+# All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are met:
+#
+# * Redistributions of source code must retain the above copyright notice, this
+#   list of conditions and the following disclaimer.
+#
+# * Redistributions in binary form must reproduce the above copyright notice,
+#   this list of conditions and the following disclaimer in the documentation
+#   and/or other materials provided with the distribution.
+#
+# * Neither the name of the copyright holder nor the names of its
+#   contributors may be used to endorse or promote products derived from
+#   this software without specific prior written permission.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+# "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+# FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+# COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+# INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+# BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+# LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+# ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+# POSSIBILITY OF SUCH DAMAGE.
+# -----------------------------------------------------------------------------
+# Authors S. Siso, STFC Daresbury Lab
+# -----------------------------------------------------------------------------
+
+''' Performs py.test tests on the Schedule PSyIR node. '''
+
+from __future__ import absolute_import
+import os
+import pytest
+from psyclone.psyir.nodes import Schedule
+from psyclone.psyGen import PSyFactory
+from psyclone.parse.algorithm import parse
+
+
+BASE_PATH = os.path.join(os.path.dirname(os.path.dirname(os.path.dirname(
+    os.path.abspath(__file__)))), "test_files", "dynamo0p3")
+
+
+def test_sched_node_str():
+    ''' Check the node_str method of the Schedule class'''
+    from psyclone.psyir.nodes.node import colored, SCHEDULE_COLOUR_MAP
+    sched = Schedule()
+    assert colored("Schedule", SCHEDULE_COLOUR_MAP["Schedule"]) in \
+        sched.node_str()
+
+
+def test_sched_getitem():
+    '''Test that Schedule has the [int] operator overloaded to return the
+    given index child'''
+    _, invoke_info = parse(os.path.join(BASE_PATH,
+                                        "15.9.1_X_innerproduct_Y_builtin.f90"),
+                           api="dynamo0.3")
+    psy = PSyFactory("dynamo0.3", distributed_memory=True).create(invoke_info)
+
+    sched = psy.invokes.invoke_list[0].schedule
+    for indx in range(len(sched._children)):
+        assert sched[indx] is sched._children[indx]
+
+    # Test range indexing
+    children = sched[:]
+    assert len(children) == 2
+    assert children[0] is sched._children[0]
+    assert children[1] is sched._children[1]
+
+    # Test index out-of-bounds Error
+    with pytest.raises(IndexError) as err:
+        _ = sched[len(sched._children)]
+    assert "list index out of range" in str(err.value)
+
+
+def test_sched_can_be_printed():
+    ''' Check the schedule class can always be printed'''
+    _, invoke_info = parse(os.path.join(BASE_PATH,
+                                        "15.9.1_X_innerproduct_Y_builtin.f90"),
+                           api="dynamo0.3")
+    psy = PSyFactory("dynamo0.3", distributed_memory=True).create(invoke_info)
+
+    # For this test use the generic class
+    psy.invokes.invoke_list[0].schedule.__class__ = Schedule
+    output = str(psy.invokes.invoke_list[0].schedule)
+
+    assert "Schedule:\n" in output
